Attempt to get the VM integrated with the scene renderer
It does not work... but it looks ok-ish. The VM is correctly executing cycles,
and if I dump the matricies they also look ok. One thing to note is the
renderer needs projection and view matricies. Because the draw callback in the
renderer VM only passes objects the projection and view matricies are not
there. So I made two special matrix registers in the VM, one for the projection
and one for the view matrix. Not sure if these should be made availabe in the
VM as they could potentially be modified by code (although this might actually
be a feature...).

+++ include/vrms_scene.h
@@ -15,7 +15,7 @@ typedef struct vrms_scene {
     uint32_t outbound_queue_index;
     pthread_mutex_t* outbound_queue_lock;
     uint32_t render_buffer_size;
-    uint32_t* render_buffer;
+    uint8_t* render_buffer;
     vrms_render_vm_t* vm;
     pthread_mutex_t* render_buffer_lock;
 } vrms_scene_t;
